+++ apps/explorer/lib/explorer/smart_contract/vyper/verifier.ex
@@ -12,7 +12,7 @@ defmodule Explorer.SmartContract.Vyper.Verifier do
   alias Explorer.Chain
   alias Explorer.SmartContract.Vyper.CodeCompiler
   alias Explorer.SmartContract.RustVerifierInterface
-  import Explorer.SmartContract.Helper, only: [prepare_bytecode_for_microservice: 3]
+  import Explorer.SmartContract.Helper, only: [prepare_bytecode_for_microservice: 3, contract_creation_input: 1]
 
   def evaluate_authenticity(_, %{"contract_source_code" => ""}),
     do: {:error, :contract_source_code}
@@ -36,16 +36,9 @@ defmodule Explorer.SmartContract.Vyper.Verifier do
       if RustVerifierInterface.enabled?() do
         deployed_bytecode = Chain.smart_contract_bytecode(address_hash)
 
-        creation_tx_input =
-          case Chain.smart_contract_creation_tx_bytecode(address_hash) do
-            %{init: init, created_contract_code: _created_contract_code} ->
-              init
+        creation_tx_input = contract_creation_input(address_hash)
 
-            _ ->
-              nil
-          end
-
-        vyper_verify_multipart(params, creation_tx_input, deployed_bytecode, params["evm_version"], files)
+        vyper_verify_multipart(params, creation_tx_input, deployed_bytecode, params["evm_version"], files, address_hash)
       end
     rescue
       exception ->
@@ -61,18 +54,18 @@ defmodule Explorer.SmartContract.Vyper.Verifier do
   defp evaluate_authenticity_inner(true, address_hash, params) do
     deployed_bytecode = Chain.smart_contract_bytecode(address_hash)
 
-    creation_tx_input =
-      case Chain.smart_contract_creation_tx_bytecode(address_hash) do
-        %{init: init, created_contract_code: _created_contract_code} ->
-          init
+    creation_tx_input = contract_creation_input(address_hash)
 
-        _ ->
-          nil
-      end
-
-    vyper_verify_multipart(params, creation_tx_input, deployed_bytecode, params["evm_version"], %{
-      "#{params["name"]}.vy" => params["contract_source_code"]
-    })
+    vyper_verify_multipart(
+      params,
+      creation_tx_input,
+      deployed_bytecode,
+      params["evm_version"],
+      %{
+        "#{params["name"]}.vy" => params["contract_source_code"]
+      },
+      address_hash
+    )
   end
 
   defp evaluate_authenticity_inner(false, address_hash, params) do
@@ -99,20 +92,14 @@ defmodule Explorer.SmartContract.Vyper.Verifier do
 
   defp compare_bytecodes({:error, _}, _, _), do: {:error, :compilation}
 
-  # credo:disable-for-next-line /Complexity/
   defp compare_bytecodes(
          {:ok, %{"abi" => abi, "bytecode" => bytecode}},
          address_hash,
          arguments_data
        ) do
     blockchain_bytecode =
-      case Chain.smart_contract_creation_tx_bytecode(address_hash) do
-        %{init: init, created_contract_code: _created_contract_code} ->
-          init
-
-        _ ->
-          nil
-      end
+      address_hash
+      |> contract_creation_input()
       |> String.trim()
 
     if String.trim(bytecode <> arguments_data) == blockchain_bytecode do
@@ -122,12 +109,12 @@ defmodule Explorer.SmartContract.Vyper.Verifier do
     end
   end
 
-  defp vyper_verify_multipart(params, creation_tx_input, deployed_bytecode, evm_version, files) do
+  defp vyper_verify_multipart(params, creation_tx_input, deployed_bytecode, evm_version, files, address_hash) do
     %{}
     |> prepare_bytecode_for_microservice(creation_tx_input, deployed_bytecode)
     |> Map.put("evmVersion", evm_version || "istanbul")
     |> Map.put("sourceFiles", files)
     |> Map.put("compilerVersion", params["compiler_version"])
-    |> RustVerifierInterface.vyper_verify_multipart()
+    |> RustVerifierInterface.vyper_verify_multipart(address_hash)
   end
 end